requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.5.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
phonenumbers==8.13.47
pillow==10.4.0
psycopg2==2.9.9
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
whitenoise==6.7.0
